requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.2
pydantic==2.5.2
python-dotenv==1.0.0 